Deleted unused files, moved miniball to img_processing
