--- loadTASinfo.m
@@ -1,142 +1,0 @@
-%function loadTASinfo(tasfilename)
-%% Import data from text file.
-% Script for importing data from the following text file:
-%
-%    /net/san-b4-ib/data/mcfarq/a/weiwu3/PECAN/01CIP20150617005743.csv
-%
-% To extend the code to different selected data or a different text file,
-% generate a function instead of a script.
-
-% Auto-generated by MATLAB on 2015/06/22 12:25:44
-
-%% Initialize variables.
-%tasfilename = '01CIP20150617005743.csv';
-
-filename = tasfilename
-filename = strsplit(filename, {'\n', ' '});
-delimiter = ',';
-startRow = 33;
-
-%% Format string for each line of text:
-
-%% Open the text file.
-[fileID, errmsg] = fopen(filename{1},'r');
-
-%% Format Spec
-formatSpec = '%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%f%[^\n\r]';
-%% Read columns of data according to format string.
-% This call is based on the structure of the file used to generate this
-% code. If an error occurs for a different file, try regenerating the code
-% from the Import Tool.
-textscan(fileID, '%[^\n\r]', startRow-1, 'ReturnOnError', false);
-dataArray = textscan(fileID, formatSpec, 'Delimiter', delimiter, 'MultipleDelimsAsOne', true, 'EmptyValue' ,NaN,'ReturnOnError', false);
-
-%% Close the text file.
-fclose(fileID);
-
-%% Post processing for unimportable data.
-% No unimportable data rules were applied during the import, so no post
-% processing code is included. To generate code which works for
-% unimportable data, select unimportable cells in a file and regenerate the
-% script.
-
-%% Allocate imported array to column variable names
-Time = dataArray{:, 1};
-Over_Rej_Count = dataArray{:, 2};
-Bin_1 = dataArray{:, 3};
-Bin_2 = dataArray{:, 4};
-Bin_3 = dataArray{:, 5};
-Bin_4 = dataArray{:, 6};
-Bin_5 = dataArray{:, 7};
-Bin_6 = dataArray{:, 8};
-Bin_7 = dataArray{:, 9};
-Bin_8 = dataArray{:, 10};
-Bin_9 = dataArray{:, 11};
-Bin_10 = dataArray{:, 12};
-Bin_11 = dataArray{:, 13};
-Bin_12 = dataArray{:, 14};
-Bin_13 = dataArray{:, 15};
-Bin_14 = dataArray{:, 16};
-Bin_15 = dataArray{:, 17};
-Bin_16 = dataArray{:, 18};
-Bin_17 = dataArray{:, 19};
-Bin_18 = dataArray{:, 20};
-Bin_19 = dataArray{:, 21};
-Bin_20 = dataArray{:, 22};
-Bin_21 = dataArray{:, 23};
-Bin_22 = dataArray{:, 24};
-Bin_23 = dataArray{:, 25};
-Bin_24 = dataArray{:, 26};
-Bin_25 = dataArray{:, 27};
-Bin_26 = dataArray{:, 28};
-Bin_27 = dataArray{:, 29};
-Bin_28 = dataArray{:, 30};
-Bin_29 = dataArray{:, 31};
-Bin_30 = dataArray{:, 32};
-Bin_31 = dataArray{:, 33};
-Bin_32 = dataArray{:, 34};
-Bin_33 = dataArray{:, 35};
-Bin_34 = dataArray{:, 36};
-Bin_35 = dataArray{:, 37};
-Bin_36 = dataArray{:, 38};
-Bin_37 = dataArray{:, 39};
-Bin_38 = dataArray{:, 40};
-Bin_39 = dataArray{:, 41};
-Bin_40 = dataArray{:, 42};
-Bin_41 = dataArray{:, 43};
-Bin_42 = dataArray{:, 44};
-Bin_43 = dataArray{:, 45};
-Bin_44 = dataArray{:, 46};
-Bin_45 = dataArray{:, 47};
-Bin_46 = dataArray{:, 48};
-Bin_47 = dataArray{:, 49};
-Bin_48 = dataArray{:, 50};
-Bin_49 = dataArray{:, 51};
-Bin_50 = dataArray{:, 52};
-Bin_51 = dataArray{:, 53};
-Bin_52 = dataArray{:, 54};
-Bin_53 = dataArray{:, 55};
-Bin_54 = dataArray{:, 56};
-Bin_55 = dataArray{:, 57};
-Bin_56 = dataArray{:, 58};
-Bin_57 = dataArray{:, 59};
-Bin_58 = dataArray{:, 60};
-Bin_59 = dataArray{:, 61};
-Bin_60 = dataArray{:, 62};
-Bin_61 = dataArray{:, 63};
-Bin_62 = dataArray{:, 64};
-DOF_Rej_Counts = dataArray{:, 65};
-End_Rej_Counts = dataArray{:, 66};
-Diode_1_Volts = dataArray{:, 67};
-Diode_64_Volts = dataArray{:, 68};
-Diode_32_Volts = dataArray{:, 69};
-Diff_Press = dataArray{:, 70};
-Static_Press = dataArray{:, 71};
-LWC_Hotwire = dataArray{:, 72};
-LWC_Slave = dataArray{:, 73};
-DSP_Board_Temp = dataArray{:, 74};
-RH = dataArray{:, 75};
-Spare_Analog_2 = dataArray{:, 76};
-Optional_Temp_1 = dataArray{:, 77};
-Optional_Temp_2 = dataArray{:, 78};
-Optional_Temp_3 = dataArray{:, 79};
-Ambient_Temp = dataArray{:, 80};
-Laser_Current = dataArray{:, 81};
-LaserPowerarbunits = dataArray{:, 82};
-Particle_Counter = dataArray{:, 83};
-Seconds = dataArray{:, 84};
-Milliseconds = dataArray{:, 85};
-Hours = dataArray{:, 86};
-Minutes = dataArray{:, 87};
-Host_Sync_Counter = dataArray{:, 88};
-Reset_Flag = dataArray{:, 89};
-True_Air_Speed = dataArray{:, 90};
-CIP_Numb = dataArray{:, 91};
-CIP_LWC = dataArray{:, 92};
-CIP_MVD = dataArray{:, 93};
-CIP_ED = dataArray{:, 94};
-Status = dataArray{:, 95};
-
-
-%% Clear temporary variables
-clearvars filename delimiter startRow formatSpec fileID dataArray ans;
